Версия 5. <feat> Добавлен классы GenreDbStorage, GenreService, MpaDbStorage, MpaService для разделения ответственности классов.           <feat> Добавлен классы FilmDbStorageTest, GenreDbStorageTest, MpaDbStorageTest, UserDbStorageTest для тестирования публичных методов Storage.           <refactor> В классе UserDbStorage добавлено использование библиотеки StringUtils.
