+++ src/main/java/ru/yandex/practicum/filmorate/controller/MpaController.java
@@ -1,14 +1,15 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.filmorate.exceptions.InvalidIdException;
-import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.Mpa;
 import ru.yandex.practicum.filmorate.service.film.FilmService;
+import ru.yandex.practicum.filmorate.service.mpa.MpaService;
 
 import java.util.Collection;
 
@@ -17,22 +18,23 @@ import java.util.Collection;
 @RequestMapping("/mpa")
 public class MpaController {
 
-    private final FilmService filmService;
+    private final MpaService mpaService;
 
-    public MpaController(FilmService filmService) {
-        this.filmService = filmService;
+    @Autowired
+    public MpaController(MpaService mpaService) {
+        this.mpaService = mpaService;
     }
 
     @GetMapping
     public Collection<Mpa> findAllMpa() {
         log.debug("Получен запрос Get /mpa. Получить все рейтинги.");
-        return filmService.findAllMpa();
+        return mpaService.findAllMpa();
     }
 
     @GetMapping("/{mpaId}")
     public Mpa getFilmById(@PathVariable int mpaId) {
         log.debug("Получен запрос Get /mpa/{}. Найти жанр по mpaId {}.", mpaId, mpaId);
-        return filmService.findMpaById(mpaId).orElseThrow(
+        return mpaService.findMpaById(mpaId).orElseThrow(
                 () -> new InvalidIdException("К сожалению, рейтинга с id " + mpaId + " нет."));
     }
 }
