+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -9,6 +9,7 @@ import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.Mpa;
 import ru.yandex.practicum.filmorate.storage.film.FilmDbStorage;
+import ru.yandex.practicum.filmorate.storage.genre.GenreDbStorage;
 import ru.yandex.practicum.filmorate.storage.user.UserDbStorage;
 import ru.yandex.practicum.filmorate.validators.FilmValidator;
 
@@ -80,21 +81,5 @@ public class FilmService {
         filmValidator.validator(film);
         log.info("Валидация в сервисе прошла успешно, объект {}", film);
         return filmStorage.updateFilm(film);
-    }
-
-    public Collection<Genre> findAllGenres() {
-        return filmStorage.findAllGenres();
-    }
-
-    public Optional<Genre> findGenreById(int genreId) {
-        return filmStorage.findGenreById(genreId);
-    }
-
-    public Collection<Mpa> findAllMpa() {
-        return filmStorage.findAllMpa();
-    }
-
-    public Optional<Mpa> findMpaById(int mpaId) {
-        return filmStorage.findMpaById(mpaId);
     }
 }
