+++ src/main/java/ru/yandex/practicum/filmorate/service/genre/GenreService.java
@@ -1,0 +1,22 @@
+package ru.yandex.practicum.filmorate.service.genre;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.model.Genre;
+import ru.yandex.practicum.filmorate.storage.genre.GenreDbStorage;
+
+import java.util.Collection;
+
+@Service
+public class GenreService {
+    private final GenreDbStorage genreDbStorage;
+
+    @Autowired
+    public GenreService(GenreDbStorage genreDbStorage) {
+        this.genreDbStorage = genreDbStorage;
+    }
+
+    public Collection<Genre> findAllGenres() {
+        return genreDbStorage.findAllGenres();
+    }
+}
