+++ src/main/java/ru/yandex/practicum/filmorate/service/mpa/MpaService.java
@@ -1,0 +1,28 @@
+package ru.yandex.practicum.filmorate.service.mpa;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.model.Mpa;
+import ru.yandex.practicum.filmorate.storage.mpa.MpaDbStorage;
+
+import java.util.Collection;
+import java.util.Optional;
+
+@Service
+public class MpaService {
+
+    private final MpaDbStorage mpaDbStorage;
+
+    @Autowired
+    public MpaService(MpaDbStorage mpaDbStorage) {
+        this.mpaDbStorage = mpaDbStorage;
+    }
+
+    public Collection<Mpa> findAllMpa() {
+        return mpaDbStorage.findAllMpa();
+    }
+
+    public Optional<Mpa> findMpaById(int mpaId) {
+        return mpaDbStorage.findMpaById(mpaId);
+    }
+}
