+++ src/main/java/ru/yandex/practicum/filmorate/storage/genre/GenreDbStorage.java
@@ -1,0 +1,52 @@
+package ru.yandex.practicum.filmorate.storage.genre;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.support.rowset.SqlRowSet;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.model.Genre;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.Collection;
+import java.util.Optional;
+
+import static ru.yandex.practicum.filmorate.constants.SqlQueryConstantsForFilm.SQL_QUERY_TAKE_ALL_GENRES;
+import static ru.yandex.practicum.filmorate.constants.SqlQueryConstantsForFilm.SQL_QUERY_TAKE_GENRE_BY_ID;
+
+@Component
+public class GenreDbStorage implements GenreStorage {
+
+    private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
+
+    @Autowired
+    public GenreDbStorage(NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
+        this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
+    }
+
+    @Override
+    public Collection<Genre> findAllGenres() {
+        return namedParameterJdbcTemplate.getJdbcTemplate()
+                .query(SQL_QUERY_TAKE_ALL_GENRES, (rs, rowNum) -> makeGenre(rs));
+    }
+
+    @Override
+    public Optional<Genre> findGenreById(int genreId) {
+        SqlRowSet genreRows = namedParameterJdbcTemplate
+                .getJdbcTemplate().queryForRowSet(SQL_QUERY_TAKE_GENRE_BY_ID, genreId);
+        if (genreRows.next()) {
+            Genre genre = new Genre(
+                    genreRows.getInt("genre_id"),
+                    genreRows.getString("genre_name"));
+            return Optional.of(genre);
+        } else {
+            return Optional.empty();
+        }
+    }
+
+    private Genre makeGenre(ResultSet rs) throws SQLException {
+        int id = rs.getInt("genre_id");
+        String name = rs.getString("genre_name");
+        return new Genre(id, name);
+    }
+}
