+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaDbStorage.java
@@ -1,0 +1,51 @@
+package ru.yandex.practicum.filmorate.storage.mpa;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.support.rowset.SqlRowSet;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.model.Mpa;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.Collection;
+import java.util.Optional;
+
+import static ru.yandex.practicum.filmorate.constants.SqlQueryConstantsForFilm.SQL_QUERY_TAKE_ALL_MPA;
+import static ru.yandex.practicum.filmorate.constants.SqlQueryConstantsForFilm.SQL_QUERY_TAKE_MPA_BY_ID;
+
+@Component
+public class MpaDbStorage implements MpaStorage {
+
+    private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
+
+    @Autowired
+    public MpaDbStorage(NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
+        this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
+    }
+
+    @Override
+    public Collection<Mpa> findAllMpa() {
+        return namedParameterJdbcTemplate.getJdbcTemplate().query(SQL_QUERY_TAKE_ALL_MPA, (rs, rowNum) -> makeMpa(rs));
+    }
+
+    @Override
+    public Optional<Mpa> findMpaById(int mpaId) {
+        SqlRowSet mpaRows = namedParameterJdbcTemplate
+                .getJdbcTemplate().queryForRowSet(SQL_QUERY_TAKE_MPA_BY_ID, mpaId);
+        if (mpaRows.next()) {
+            Mpa mpa = new Mpa(
+                    mpaRows.getInt("mpa_id_in_mpa"),
+                    mpaRows.getString("mpa_name"));
+            return Optional.of(mpa);
+        } else {
+            return Optional.empty();
+        }
+    }
+
+    private Mpa makeMpa(ResultSet rs) throws SQLException {
+        int id = rs.getInt("mpa_id_in_mpa");
+        String name = rs.getString("mpa_name");
+        return new Mpa(id, name);
+    }
+}
