+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaStorage.java
@@ -1,0 +1,13 @@
+package ru.yandex.practicum.filmorate.storage.mpa;
+
+import ru.yandex.practicum.filmorate.model.Mpa;
+
+import java.util.Collection;
+import java.util.Optional;
+
+public interface MpaStorage {
+
+    Collection<Mpa> findAllMpa();
+
+    Optional<Mpa> findMpaById(int mpaId);
+}
